refactor(api)!: move `Connector` enum to `net` mod
- refactor the implementation of methods
  `fedimint-api-client::api::download_from_invite_code` and
  `try_download_client_config` for `Connector` enum instead.
- add usage of the methods implemented for `Connector` to reduce the
  use/pass of `connector` parameter all around.
