+++ fedimint-api-client/src/lib.rs
@@ -6,7 +6,8 @@
 #![allow(clippy::return_self_not_must_use)]
 
 use anyhow::{bail, Context as _};
-use api::{Connector, DynGlobalApi, FederationApiExt as _, WsFederationApi};
+use api::net::Connector;
+use api::{DynGlobalApi, FederationApiExt as _, WsFederationApi};
 use fedimint_core::config::{ClientConfig, FederationId};
 use fedimint_core::endpoint_constants::CLIENT_CONFIG_ENDPOINT;
 use fedimint_core::invite_code::InviteCode;
@@ -20,71 +21,72 @@ pub mod api;
 /// Client query system
 pub mod query;
 
-// FIXME: (@leonardo) how should this fns handle different [`Connector`]'s
+impl Connector {
+    /// Tries to download the [`ClientConfig`] from the federation with an
+    /// specified [`Connector`] variant, attempts to retry ten times before
+    /// giving up.
+    pub async fn download_from_invite_code(
+        &self,
+        invite_code: &InviteCode,
+    ) -> anyhow::Result<ClientConfig> {
+        debug!("Downloading client config from {:?}", invite_code);
 
-/// Tries to download the [`ClientConfig`] from the federation,
-/// attempts to retry ten times before giving up.
-pub async fn download_from_invite_code(
-    connector: Connector,
-    invite_code: &InviteCode,
-) -> anyhow::Result<ClientConfig> {
-    debug!("Downloading client config from {:?}", invite_code);
+        let federation_id = invite_code.federation_id();
+        // FIXME: (@leonardo) should fetch all the api_endpoints with proper
+        // [`Connector`] too!
+        let api = DynGlobalApi::from_invite_code(invite_code);
+        let api_secret = invite_code.api_secret();
 
-    let federation_id = invite_code.federation_id();
-    // FIXME: (@leonardo) should fetch all the api_endpoints with proper
-    // [`Connector`] too!
-    let api = DynGlobalApi::from_invite_code(invite_code);
-    let api_secret = invite_code.api_secret();
-
-    fedimint_core::util::retry(
-        "Downloading client config",
-        backoff_util::aggressive_backoff(),
-        || try_download_client_config(&api, connector, federation_id, api_secret.clone()),
-    )
-    .await
-    .context("Failed to download client config")
-}
-
-/// Tries to download the client config only once.
-pub async fn try_download_client_config(
-    api: &DynGlobalApi,
-    connector: Connector,
-    federation_id: FederationId,
-    api_secret: Option<String>,
-) -> anyhow::Result<ClientConfig> {
-    // TODO: use new download approach based on guardian PKs
-    let query_strategy = FilterMap::new(
-        move |cfg: ClientConfig| {
-            if federation_id != cfg.global.calculate_federation_id() {
-                bail!("FederationId in invite code does not match client config")
-            }
-
-            Ok(cfg.global.api_endpoints)
-        },
-        NumPeers::from(1),
-    );
-
-    let api_endpoints = api
-        .request_with_strategy(
-            query_strategy,
-            CLIENT_CONFIG_ENDPOINT.to_owned(),
-            ApiRequestErased::default(),
+        fedimint_core::util::retry(
+            "Downloading client config",
+            backoff_util::aggressive_backoff(),
+            || self.try_download_client_config(&api, federation_id, api_secret.clone()),
         )
-        .await?;
-
-    // now we can build an api for all guardians and download the client config
-    let api_endpoints = api_endpoints.into_iter().map(|(peer, url)| (peer, url.url));
-
-    let client_config = WsFederationApi::new(&connector, api_endpoints, &api_secret)
-        .request_current_consensus::<ClientConfig>(
-            CLIENT_CONFIG_ENDPOINT.to_owned(),
-            ApiRequestErased::default(),
-        )
-        .await?;
-
-    if client_config.calculate_federation_id() != federation_id {
-        bail!("Obtained client config has different federation id");
+        .await
+        .context("Failed to download client config")
     }
 
-    Ok(client_config)
+    /// Tries to download the [`ClientConfig`] only once.
+    pub async fn try_download_client_config(
+        &self,
+        api: &DynGlobalApi,
+        federation_id: FederationId,
+        api_secret: Option<String>,
+    ) -> anyhow::Result<ClientConfig> {
+        // TODO: use new download approach based on guardian PKs
+        let query_strategy = FilterMap::new(
+            move |cfg: ClientConfig| {
+                if federation_id != cfg.global.calculate_federation_id() {
+                    bail!("FederationId in invite code does not match client config")
+                }
+
+                Ok(cfg.global.api_endpoints)
+            },
+            NumPeers::from(1),
+        );
+
+        let api_endpoints = api
+            .request_with_strategy(
+                query_strategy,
+                CLIENT_CONFIG_ENDPOINT.to_owned(),
+                ApiRequestErased::default(),
+            )
+            .await?;
+
+        // now we can build an api for all guardians and download the client config
+        let api_endpoints = api_endpoints.into_iter().map(|(peer, url)| (peer, url.url));
+
+        let client_config = WsFederationApi::new(self, api_endpoints, &api_secret)
+            .request_current_consensus::<ClientConfig>(
+                CLIENT_CONFIG_ENDPOINT.to_owned(),
+                ApiRequestErased::default(),
+            )
+            .await?;
+
+        if client_config.calculate_federation_id() != federation_id {
+            bail!("Obtained client config has different federation id");
+        }
+
+        Ok(client_config)
+    }
 }
