+++ fedimint-server/src/consensus/mod.rs
@@ -15,7 +15,8 @@ use std::time::Duration;
 use anyhow::bail;
 use async_channel::Sender;
 use db::{get_global_database_migrations, GLOBAL_DATABASE_VERSION};
-use fedimint_api_client::api::{Connector, DynGlobalApi};
+use fedimint_api_client::api::net::Connector;
+use fedimint_api_client::api::DynGlobalApi;
 use fedimint_core::config::ServerModuleInitRegistry;
 use fedimint_core::core::{ModuleInstanceId, ModuleKind};
 use fedimint_core::db::{apply_migrations, apply_migrations_server, Database};
