+++ fedimint-wasm-tests/src/lib.rs
@@ -37,11 +37,9 @@ async fn make_client_builder() -> Result<fedimint_client::ClientBuilder> {
 }
 
 async fn client(invite_code: &InviteCode) -> Result<fedimint_client::ClientHandleArc> {
-    let client_config = fedimint_api_client::download_from_invite_code(
-        fedimint_api_client::api::Connector::default(),
-        invite_code,
-    )
-    .await?;
+    let client_config = fedimint_api_client::api::net::Connector::default()
+        .download_from_invite_code(invite_code)
+        .await?;
     let mut builder = make_client_builder().await?;
     let client_secret = load_or_generate_mnemonic(builder.db_no_decoders()).await?;
     builder.stopped();
